Issue#104 (#112)
* fixed issue #106
hide password on registration page

* fixed issue #103
hide password while typing on index page

* fixed issue #104

Price of products accepts floats now

+++ src/main/java/team24soft/minishopify/controllers/ShopController.java
@@ -59,7 +59,7 @@ public class ShopController {
     public String getItems(@RequestParam(name="id", required = true) long id,
                            @RequestParam(name="name", required = true) String name,
                            @RequestParam(name="description", required = false,defaultValue = "") String  description,
-                           @RequestParam(name="price", required = false, defaultValue = "0") long price,
+                           @RequestParam(name="price", required = false, defaultValue = "0") float price,
                            @RequestParam(name="quantity", required = false, defaultValue = "0") long quantity,
                            @RequestParam(name="userId", required = true) long userId,
                            Model model){
